Divide Prepare Dataset Step in two components
