+++ DashAI/front/src/components/experiments/ConfigureModelsStep.jsx
@@ -160,9 +160,9 @@ ConfigureModelsStep.propTypes = {
     input_columns: PropTypes.arrayOf(PropTypes.number),
     output_columns: PropTypes.arrayOf(PropTypes.number),
     splits: PropTypes.shape({
-      train: PropTypes.number,
+      training: PropTypes.number,
       validation: PropTypes.number,
-      test: PropTypes.number,
+      testing: PropTypes.number,
     }),
     step: PropTypes.string,
     created: PropTypes.instanceOf(Date),
